Relocate tools package into Overlay plugin system
