+++ Overlay/plugins/tools/tool/list.py
@@ -1,10 +1,12 @@
 import json
 import sys
+from pathlib import Path
 from .registry import ToolRegistry
 from .base import ToolContext
 
 def main():
-    yaml_path = sys.argv[1] if len(sys.argv) > 1 else "tools/config/tools.yaml"
+    default_yaml = Path(__file__).resolve().parents[2] / "config" / "tools.yaml"
+    yaml_path = sys.argv[1] if len(sys.argv) > 1 else str(default_yaml)
     reg = ToolRegistry.from_yaml(yaml_path, ctx=ToolContext())
     print(json.dumps(reg.list_tools(), ensure_ascii=False, indent=2))
